Linear Operators

JUDI is building on JOLI.jl to implement matrix-free linear operators. These operators represent the discretized wave-equations and sensitivit (Jacobian) for different acquisition schemes.

Linear Operators

judiModeling

judiJacobian

judiProjection

judiLRWF

judiModeling

Seismic modeling operator for solving a wave equation for a given right-hand-side.

JUDI.judiModeling — Type

judiModeling(info, model; options=Options())
judiModeling(model, src_geometry, rec_geometry; options=Options())
judiModeling(info, model, src_geometry, rec_geometry; options=Options())

Create seismic modeling operator for a velocity model given as a Model structure. info is an Info structure
containing necessary dimensions to set up the operator. The function also takes the source and receiver geometries
as additional input arguments, which creates a combined operator judiProjection*judiModeling*judiProjection'.

Example

Pr and Ps are projection operatos of type judiProjection and
q is a data vector of type judiVector:

F = judiModeling(info, model)
dobs = Pr*F*Ps'*q

F = judiModeling(info, model, q.geometry, rec_geometry)
dobs = F*q

Construction:

Construct a modeling operator without source/receiver projections:

F = judiModeling(info, model; options=opt)

Construct a modeling operator with source/receiver projections:

F = judiModeling(info, model, src_geometry, rec_geometry)

Not that in this case, all the information can be infered from the model and geometries. Therefore the operator can be created without the need for the info input:

F = judiModeling(model, src_geometry, rec_geometry)

Construct a modeling operator from an existing operator without geometries and projection operators:

F = Pr*F*Ps'

where Ps and Pr are source/receiver projection operators of type judiProjection.

Construct a modeling operator for extended source modeling:

F = Pr*F*Pw'

where Pw is a judiLRWF (low-rank-wavefield) projection operator.

JUDI.judiModeling — Method

judiModeling(info, model; options=Options())
judiModeling(model, src_geometry, rec_geometry; options=Options())
judiModeling(info, model, src_geometry, rec_geometry; options=Options())

Create seismic modeling operator for a velocity model given as a Model structure. info is an Info structure
containing necessary dimensions to set up the operator. The function also takes the source and receiver geometries
as additional input arguments, which creates a combined operator judiProjection*judiModeling*judiProjection'.

Example

Pr and Ps are projection operatos of type judiProjection and
q is a data vector of type judiVector:

F = judiModeling(info, model)
dobs = Pr*F*Ps'*q

F = judiModeling(info, model, q.geometry, rec_geometry)
dobs = F*q

Accessible fields:

# Info structure
F.info

# Model structure
F.model

# Source geometry (if available)
F.srcGeometry

# Receiver geometry (if available)
F.recGeometry

# Options structure
F.options

Usage:

# Forward modeling (F w/ geometries)
d_obs = F*q

# Adjoint modeling (F w/ geometries)
q_ad = F'*d_obs

# Forward modeling (F w/o geometries)
d_obs = Pr*F*Ps'*q

# Adjoint modelng (F w/o geometries)
q_ad = Ps*F'*Pr'*d_obs

# Extended source modeling (F w/o geometries)
d_obs  = Pr*F*Pw'*w

# Adjoint extended source modeling (F w/o geometries)
w_ad = Pw*F'*Pr'*d_obs

# Forward modeling and return full wavefield (F w/o geometries)
u = F*Ps'*q

# Adjoint modelnig and return wavefield (F w/o geometries)
v = F'*Pr'*d_obs

# Forward modeling with full wavefield as source (F w/o geometries)
d_obs = Pr*F*u

# Adjoint modeling with full wavefield as source (F w/o geometries)
q_ad = Ps*F*v

judiJacobian

Jacobian of a non-linear forward modeling operator. Corresponds to linearized Born modeling (forward mode) and reverse-time migration (adjoint mode).

JUDI.judiJacobian — Type

judiJacobian(F,q)

Create a linearized modeling operator from the non-linear modeling operator F and
the source q. F is a full modeling operator including source/receiver projections.

Examples

F is a modeling operator without source/receiver projections:

J = judiJacobian(PrFPs',q)

F is the combined operator Pr*F*Ps':

J = judiJacobian(F,q)

judiJacobianExQ(F,q)

Create a linearized modeling operator from the non-linear modeling operator F and
the source q. F is a full modeling operator including source/receiver projections. Examples ========

F is a modeling operator without source/receiver projections: J = judiJacobianExQ(PrFPs',q)

F is the combined operator Pr*F*Ps': J = judiJacobianExQ(F,q)

Construction:

A judiJacobian operator can be create from an exisiting forward modeling operator and a source vector:

J = judiJacobian(F, q)  # F w/ geometries

J = judiJacobian(Pr*F*Ps', q)   # F w/o geometries

where Ps and Pr are source/receiver projection operators of type judiProjection.

A Jacobian can also be created for an extended source modeling operator:

J = judiJacobian(Pr*F*Pw', w)

where Pw is a judiLRWF operator and w is a judiWeights vector (or 2D/3D Julia array).

Accessible fields::

# Info structure
J.info

# Model structure
J.model

# Source geometry (if available)
J.srcGeometry

# Receiver geometry
J.recGeometry

# Source wavelet
J.wavelet

# Weights (extended source modeling only)
J.weights

# Options structure
J.options

Usage:

# Linearized modeilng
d_lin = J*dm

# RTM
rtm = J'*d_lin

# Matrix-free normal operator
H = J'*J

judiProjection

Abstract linear operator for source/receiver projections. A (transposed) judiProjection operator symbolically injects the data with which it is multiplied during modeling. If multiplied with a forward modeling operator, it samples the wavefield at the specified source/receiver locations.

JUDI.judiProjection — Type

judiProjection(info, geometry)

Projection operator for sources/receivers to restrict or inject data at specified locations.
info is an Info structure and geometry is a Geometry structure with either source or
receiver locations.

Examples

F is a modeling operator of type judiModeling and q is a seismic source of type judiVector:

Pr = judiProjection(info, rec_geometry)
Ps = judiProjection(info, q.geometry)

dobs = Pr*F*Ps'*q
qad = Ps*F'*Pr'*dobs

Accessible fields:

# Info structure
P.info

# Source/receiver geometry
P.geometry

Usage:

# Multiply with judiVector to create a judiRHS
rhs1 = Pr'*d_obs
rhs2 = Ps'*q

# Sample wavefield at source/receiver location during modeling
d_obs = Pr*F*Ps'*q
q_ad = Ps*F*Pr'*d_obs

judiLRWF

Abstract linear operator for sampling a seismic wavefield as a sum over all time steps, weighted by a time-varying wavelet. Its transpose injects a time-varying wavelet at every grid point in the model.

JUDI.judiLRWF — Type

judiLRWF(info, geometry)

Low-rank wavefield operator which injects a wavelet q at every point of the subsurface.
info is an Info structure and wavelet is a cell array containing the wavelet(s). Examples ======== F is a modeling operator of type judiModeling and w is a weighting matrix of type judiWeights: Pr = judiProjection(info, rec_geometry) Pw = judiLRWF(info, q.data) dobs = PrFPw'w dw = PwF'Pr'dobs

Accessible fields:

# Info structure
P.info

# Wavelet of i-th source location
P.wavelet[i]

Usage:

# Multiply with a judiWeight vector to create a judiExtendedSource
ex_src = Pw'*w

# Sample wavefield as a sum over time, weighted by the source
u_ex = Pw*F'*Pr'*d_obs